Move ExternalDependency to a new file
